Firebase로부터 데이터를 불러오는 메서드 분리
