+++ RandomStudy/SQLite/DBHelper.swift
@@ -29,8 +29,6 @@ class DBHelper {
             print("create table \(i)")
             print("table \(i): \(self.readData(tableName: i, column: column))")
         }
-        
-        
     }
     
     deinit {
@@ -222,95 +220,12 @@ class DBHelper {
         self.delegate?.removeAllDatas()
     }
     
-//    MARK: Data Migration (SQLite3 -> Firebase)
-    func isDataExist() -> Bool{
+    func makeSampleData() {
         for i in tableNames {
-            let data = DBHelper.shared.readData(tableName: i, column:  column)
-            if !data.isEmpty { return true }
+            self.insertData(tableName: i, columns: column, insertData: ["\(i)_1","0","0"])
+            print("DBHelper:: makeSampleData:: \(self.readData(tableName: i, column: column))")
         }
-        return false
-    }
-    func getUserInfo(completion: @escaping (String?) -> ()) {
-        Auth.auth().addStateDidChangeListener { auth, user in
-            completion(user?.uid)
-        }
-    }
-    func dataMigration() {
-        self.getUserInfo { [weak self] uid in
-            guard let uid = uid else { return }
-            guard let tables = self?.tableNames else { return }
-            guard let column = self?.column else { return }
-            guard let self = self else { return }
-            do {
-                try Firestore.firestore().collection("users").document(uid).setData(["uid": uid])
-                print("success written uid")
-            } catch {
-                print("fail writing uid")
-            }
-            for i in tables {
-                let data = self.readData(tableName: i, column: column)
-                if data.isEmpty { continue }
-                print(data)
-                var dataArray = [[String: String]]()
-                for j in data {
-                    dataArray.append(["name": j.name!, "done": j.done!, "date": j.date!])
-                }
-                do {
-                    try Firestore.firestore().collection("users").document(uid).updateData(["\(i)": dataArray])
-                    print("success written document")
-                } catch {
-                    print("fail writing document")
-                }
-                
-            }
-            self.resetAllTable()
-        }
-    }
-    func getDataFromFirebase(dataName: String, completion: @escaping ([StudyModel]?) -> ()) {
-        guard let uid = Auth.auth().currentUser?.uid else {
-            completion(nil)
-            return
-        }
-        let db = Firestore.firestore()
-        db.collection("users").whereField("uid", isEqualTo: uid).getDocuments { snapshot, error in
-            if error != nil {
-                let networkError = error as! NSError
-                print(networkError.localizedDescription)
-                completion(nil)
-                return
-            }
-            if snapshot!.documents.isEmpty {
-                completion([])
-                return
-            }
-            
-            guard let filtered = snapshot?.documents[0].data().filter({ (key: String, value: Any) in
-                if key == dataName { return true }
-                return false
-            }) else { return }
-            
-            do {
-                let jsonFile = try JSONSerialization.data(withJSONObject: filtered)
-                let data = try JSONDecoder().decode([String: [StudyModel]].self, from: jsonFile)
-                let converted = Array(data.values).flatMap{$0}
-                completion(converted)
-                print("receive Data Successful")
-            } catch {
-                print("receive Data Fail")
-            }
-        }
-    }
-    func removeFirebaseData() {
-        guard let uid = Auth.auth().currentUser?.uid else { return }
-        let db = Firestore.firestore()
-        do {
-            try db.collection("users").document(uid).updateData(["todo": FieldValue.delete()])
-            try db.collection("users").document(uid).updateData(["history": FieldValue.delete()])
-            try db.collection("users").document(uid).updateData(["study": FieldValue.delete()])
-            print("DBHelper:: delete Success")
-        } catch {
-            print("DBHelper:: do delete fail")
-        }
+        
     }
 }
